backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database & ORM
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
httpx==0.27.2
google-auth==2.35.0
authlib==1.3.2
pyopenssl==24.2.1

# AI & LangChain
langchain==1.2.3
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-community==0.4.1
openai==2.15.0

# Cache & Utilities
redis==5.2.0
python-dotenv==1.0.1

# Security & Rate Limiting
slowapi==0.1.9
itsdangerous==2.2.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
